"{#8} {product.php} {http://pm.5psolutions.net/public/index.php?path_info=projects/training-andrei-ghiorghe/tasks/8} {This commit updated product.php in order to respect the specs of the project}"

+++ product.php
@@ -49,17 +49,35 @@ if (isset($_POST['editButton'])) {
 
 if (isset($_POST['submitButton'])) {
     // validate fields: title, description and price
-    $data = validateFields(['inputData' => $inputData, 'post' => $_POST]);
-    $inputData = $data['inputData'];
-    $inputError = $data['inputError'];
-
-    if (isset($_FILES['imageFileField']) && $_FILES['imageFileField']['tmp_name']) {
-        $inputData['imageLocation'] = $_FILES['imageFileField']['tmp_name'];
-        $inputData['imageName'] = $_FILES['imageFileField']['name'];
+    if (isset($_POST['titleField']) && $_POST['titleField']) {
+        $inputData['titleField'] = $_POST['titleField'];
     } else {
+        $inputError['titleFieldError'] = 'Please enter a title for product';
+    }
+    if (isset($_POST['descriptionField']) && $_POST['descriptionField']) {
+        $inputData['descriptionField'] = $_POST['descriptionField'];
+    } else {
+        $inputError['descriptionFieldError'] = 'Please enter a description for product';
+    }
+    if (isset($_POST['priceField'])) {
+        $inputData['priceField'] = $_POST['priceField'];
+
+        if (!is_numeric($_POST['priceField']) || !intval($_POST['priceField'])) {
+            $inputError['priceFieldError'] = 'Please enter a natural number as price for product';
+        }
+    }
+
+    // validate the files input
+    if (isset($_FILES['fileField']) && $_FILES['fileField']['tmp_name']) {
+        $inputData['imageLocation'] = $_FILES['fileField']['tmp_name'];
+        $inputData['imageName'] = $_FILES['fileField']['name'];
+    } elseif (!(isset($oldImagePath) && $oldImagePath)) {
+        // if I don't have something in $_FILES and I don't have an old image set, then
+        // it means I have no image for the product
         $inputError['imageFileFieldError'] = 'Please choose an image for the product';
     }
 
+    // if I have no errors util now
     if (!count($inputError)) {
         // idProductEdit is equal 0 when there is no product for editing
         if (!$idProductEdit) {
@@ -78,33 +96,50 @@ if (isset($_POST['submitButton'])) {
             header('Location: products.php');
             die();
         } else {
-            $imagePath = 'images/' . time() . $inputData['imageName'];
-            $response = query(
-                $connection,
-                'UPDATE products SET title = ?, description = ?, price = ?, image_path = ? WHERE id = ?',
-                [
+            // check if we don't have the old image
+            // if I do not have it that means I have to replace it with the new one
+            if (!(isset($oldImagePath) && $oldImagePath)) {
+                $queryString = 'UPDATE products 
+                                SET title = ?, description = ?, price = ?, image_path = ? 
+                                WHERE id = ?';
+                $imagePath = 'images/' . time() . $inputData['imageName'];
+                $paramQuery = [
                     $inputData['titleField'],
                     $inputData['descriptionField'],
                     intval($inputData['priceField']),
                     $imagePath,
                     $idProductEdit,
-                ]
+                ];
+            } else {
+                $queryString = 'UPDATE products SET title = ?, description = ?, price = ? WHERE id = ?';
+                $paramQuery = [
+                    $inputData['titleField'],
+                    $inputData['descriptionField'],
+                    intval($inputData['priceField']),
+                    $idProductEdit,
+                ];
+            }
+
+            // the execution of the query
+            $response = query(
+                $connection,
+                $queryString,
+                $paramQuery
             );
-            move_uploaded_file($inputData['imageLocation'], $imagePath);
-            // delete the old image\
-            if (file_exists($oldImagePath)) {
-                unlink($oldImagePath);
+
+            // when the admin chooses a new image we have to delete the old one
+            if (!(isset($oldImagePath) && $oldImagePath)) {
+                move_uploaded_file($inputData['imageLocation'], $imagePath);
+                // delete the old image\
+                if (file_exists($oldImagePath)) {
+                    unlink($oldImagePath);
+                }
             }
             header('Location: products.php');
             die();
         }
     }
 }
-
-
-
-var_dump($_POST);
-var_dump($_GET);
 ?>
 <!DOCTYPE html>
 <html>
@@ -112,21 +147,21 @@ var_dump($_GET);
     <link rel="stylesheet" type="text/css" href="http://localhost/Test/frontend/style.css">
     <meta charset="UTF-8">
     <title> <?= translate('Admin\'s page') ?> </title>
+    <script>
+        // I need this function when I want to save the name of the image
+        // and tha function it changes the default value of the label to
+        // the actual value of the input file
+        function changeLabel() {
+            var asta = document.getElementById('inputFileId').value
+            var aux = asta.split('\\')[asta.split('\\').length - 1]
+            document.getElementById('labelId').innerText = aux;
+        }
+    </script>
 </head>
-<body>
+<body onload="LoadValue();">
     <table id="contentTable">
         <tbody>
             <form action="product.php" method="POST" enctype="multipart/form-data">
-                <input
-                        type="hidden"
-                        name="idProductEdit"
-                        value="<?= isset($idProductEdit) ? $idProductEdit : '' ?>"
-                >
-                <input
-                        type="hidden"
-                        name="oldImagePath"
-                        value="<?= isset($oldImagePath) ? $oldImagePath : '' ?>"
-                >
                 <tr>
                     <td>
                         <input
@@ -172,25 +207,38 @@ var_dump($_GET);
                         </span>
                     </td>
                 </tr>
-<!--                <tr>-->
-<!--                    <td>-->
-<!--                        <input type="file" name="imageFileField">-->
-<!--                        <span class="errorField">-->
-<!--                            --><?//= isset($inputError['imageFileFieldError'])
-//                                ? translate($inputError['imageFileFieldError'])
-//                                : '' ?>
-<!--                        </span>-->
-<!--                    </td>-->
-<!--                </tr>-->
-                <?php var_dump($_FILES) ?>
                 <tr>
-                    <td><button onclick="this.children[0].click()" name="button" >
-                            <?php var_dump($_FILES) ?>
-                            <input type="file" name="imageFileFieldd" style="display: none;">
-
-
-                        </button><span class="errorField">
-                                                    </span></td>
+                    <td>
+                        <input
+                                type="hidden"
+                                name="idProductEdit"
+                                value="<?= isset($idProductEdit) ? $idProductEdit : '' ?>"
+                        >
+                        <input
+                                type="hidden"
+                                name="oldImagePath"
+                                value="<?= isset($oldImagePath) ? $oldImagePath : '' ?>"
+                        >
+                        <label for="inputFileId" id="labelId" name="labelId">
+                                <?php
+                                if (isset($oldImagePath) && $oldImagePath) {
+                                    $string = explode('/', $oldImagePath);
+                                    $string = ($string) ? $string : [translate('Error: Please Choose a Image')];
+                                    echo $string[count($string) - 1];
+                                } else {
+                                    echo 'Choose an Image: Click Here!';
+                                }
+                                ?>
+                        </label>
+                        <input onchange="changeLabel()" type="file" id="inputFileId" style="display:none" name="fileField">
+                        <span class="errorField">
+                            <?= isset($inputError['imageFileFieldError'])
+                                ? translate($inputError['imageFileFieldError'])
+                                : '' ?>
+                        </span>
+                    </td>
+                </tr>
+                <tr>
                     <td>
                         <button type="submit" name="goToProducts" class="linkButton">
                             <?= translate('products') ?>
